# openapi.yaml
openapi: 3.1.0
info:
  title: OssPilot Tenant API
  version: 0.1.0
paths:
...
  /api/buckets/{bucket_name}/policy:
    get:
      summary: Get S3 bucket policy
      responses:
        "200":
          description: policy or empty
    put:
      summary: Replace S3 bucket policy
      responses:
        "200":
          description: saved
        "400":
          description: invalid policy
    delete:
      summary: Delete S3 bucket policy
      responses:
        "204":
          description: deleted
  /api/buckets/{bucket_name}/cors:
    get:
      summary: Get bucket CORS rules
      responses:
        "200":
          description: rules
    put:
      summary: Replace bucket CORS rules
      responses:
        "200":
          description: saved
        "400":
          description: invalid rules
    delete:
      summary: Delete bucket CORS
      responses:
        "204":
          description: deleted
...
  /internal/accounts/{username}/buckets:
    put:
      summary: Replace ops-sourced bucket grants
      responses:
        "204":
          description: replaced
        "404":
          description: account missing
...
  /api/editor/open:
    post:
      summary: Open an ONLYOFFICE session for docx/xlsx/pptx
      responses:
        "200":
          description: office_url and config
        "503":
          description: OFFICE_URL unset
...
  /api/editor/save:
    post:
      summary: Latest saved metadata for an office session
      responses:
        "200":
          description: saved
        "409":
          description: callback has not saved yet
...